_posts/assets/0404_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.10.0
filelock==3.18.0
flake8==7.2.0
flake8-bugbear==24.12.12
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.3.2
networkx==3.4.2
numpy==2.2.4
openai==1.70.0
packaging==24.2
propcache==0.3.1
pycodestyle==2.13.0
pydantic==2.11.2
pydantic_core==2.33.1
pyflakes==3.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.18.3
